Instance detail model refactor
