+++ src/module/data/instance-detail/instance-detail.component.ts
@@ -4,63 +4,22 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-import { Component, Input, OnInit, OnDestroy } from "@angular/core";
+import { Component, inject, Input, OnInit } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { MatProgressSpinnerModule } from "@angular/material/progress-spinner";
 import { MatButtonModule } from "@angular/material/button";
 import { MatTooltip, MatTooltipModule } from "@angular/material/tooltip";
 import { Clipboard } from "@angular/cdk/clipboard";
-import { Subscription } from "rxjs";
-import { filter, take } from "rxjs/operators";
-import { SchemaAttribute, SchemaConcept, SchemaState } from "../../../service/schema-state.service";
-import { DriverState } from "../../../service/driver-state.service";
+import { SchemaConcept, SchemaState } from "../../../service/schema-state.service";
 import { SnackbarService } from "../../../service/snackbar.service";
 import { BreadcrumbItem, DataEditorState } from "../../../service/data-editor-state.service";
-import { ApiResponse, Concept, ConceptRowAnswer, Entity, isApiErrorResponse, QueryResponse, Relation, RoleType } from "@typedb/driver-http";
-import { extractErrorMessage } from "../../../framework/util/observable";
-
-function isInstance(concept: Concept | undefined): concept is Entity | Relation {
-    return concept?.kind === "entity" || concept?.kind === "relation";
-}
-
-function isRoleType(concept: Concept | undefined): concept is RoleType {
-    return concept?.kind === "roleType";
-}
-
-interface AttributeData {
-    type: string;
-    valueType: string;
-    values: string[];
-}
-
-interface RelationInstanceData {
-    relationTypeLabel: string;
-    relationIID: string;
-    links: LinkData[];
-    attributes: AttributeData[];
-    attributesLoaded: boolean;
-}
-
-interface LinkData {
-    roleLabel: string;
-    playerIID: string;
-    playerTypeLabel: string;
-    isSelf: boolean;
-    attributes: AttributeData[];
-    expanded: boolean;
-}
-
-interface OwnerData {
-    ownerIID: string;
-    ownerTypeLabel: string;
-    ownerKind: "entity" | "relation";
-    attributes: AttributeData[];
-}
+import { InstanceDetailViewModel, LinkData, OwnerData, RelationInstanceData } from "../../../service/instance-detail-view-model.service";
 
 @Component({
     selector: "ts-instance-detail",
     templateUrl: "./instance-detail.component.html",
     styleUrls: ["./instance-detail.component.scss"],
+    providers: [InstanceDetailViewModel],
     imports: [
         CommonModule,
         MatProgressSpinnerModule,
@@ -68,59 +27,37 @@ interface OwnerData {
         MatTooltipModule,
     ],
 })
-export class InstanceDetailComponent implements OnInit, OnDestroy {
+export class InstanceDetailComponent implements OnInit {
     @Input({ required: true }) type!: SchemaConcept;
     @Input({ required: true }) instanceIID!: string;
     @Input() breadcrumbs: BreadcrumbItem[] = [];
 
-    attributes: AttributeData[] = [];
-    allRelations: RelationInstanceData[] = [];
-    relationTypes: string[] = [];
+    vm = inject(InstanceDetailViewModel);
+
+    // Per-component UI state
     selectedRelationType: string | null = null; // null = "All"
-    loading = false;
-    relationsLoading = false;
-    linksLoading = false;
-    ownLinks: LinkData[] = [];
-    owners: OwnerData[] = [];
-    ownersLoading = false;
     typeCollapsed = false;
     linksCollapsed = false;
     attributesCollapsed = false;
     relationsCollapsed = false;
     ownersCollapsed = false;
 
-    /** True when in manual mode with no open transaction */
-    needsTransaction = false;
+    private snackbar = inject(SnackbarService);
+    private clipboard = inject(Clipboard);
+    private schemaState = inject(SchemaState);
+    private dataEditorState = inject(DataEditorState);
 
-    /** Transaction ID used to load the current data (for stale detection) */
-    private loadedWithTransactionId: string | null = null;
-
-    /** True when data was loaded in a different transaction than the current one */
-    isDataStale = false;
-
-    private subscriptions: Subscription[] = [];
-
-    constructor(
-        private driver: DriverState,
-        private snackbar: SnackbarService,
-        private clipboard: Clipboard,
-        private schemaState: SchemaState,
-        private dataEditorState: DataEditorState,
-    ) {}
+    ngOnInit() {
+        this.vm.initialize(this.type, this.instanceIID);
+    }
 
     get filteredRelations(): RelationInstanceData[] {
-        if (this.selectedRelationType === null) {
-            return this.allRelations;
-        }
-        return this.allRelations.filter(r => r.relationTypeLabel === this.selectedRelationType);
+        if (this.selectedRelationType === null) return this.vm.allRelations;
+        return this.vm.allRelations.filter(r => r.relationTypeLabel === this.selectedRelationType);
     }
 
     selectRelationType(type: string | null) {
         this.selectedRelationType = type;
-    }
-
-    getRelationTypeCount(type: string): number {
-        return this.allRelations.filter(r => r.relationTypeLabel === type).length;
     }
 
     copyValue(value: string, tooltip: MatTooltip) {
@@ -129,432 +66,42 @@ export class InstanceDetailComponent implements OnInit, OnDestroy {
         setTimeout(() => tooltip.hide(0), 1000);
     }
 
-    ngOnInit() {
-        // Check if we need a transaction before loading data
-        if (this.checkNeedsTransaction()) {
-            // Subscribe to transaction state changes to auto-load when transaction opens
-            this.subscriptions.push(
-                this.driver.transaction$.pipe(
-                    filter(tx => tx != null),
-                    take(1) // Only react once when transaction first opens
-                ).subscribe(() => {
-                    this.needsTransaction = false;
-                    this.loadData();
-                })
-            );
-        } else {
-            this.loadData();
-        }
-
-        // Subscribe to transaction changes to detect stale data (only relevant in manual mode)
-        this.subscriptions.push(
-            this.driver.transaction$.subscribe(tx => {
-                // Only show stale warning in manual mode when a different transaction is open
-                if (!this.driver.autoTransactionEnabled$.value && tx != null && tx.id !== this.loadedWithTransactionId) {
-                    this.isDataStale = true;
-                }
-                // Clear stale flag when switching to auto mode or when transaction is closed
-                if (this.driver.autoTransactionEnabled$.value || tx == null) {
-                    this.isDataStale = false;
-                }
-            })
-        );
-
-        // Subscribe to auto mode changes to load data when switching from manual to auto
-        this.subscriptions.push(
-            this.driver.autoTransactionEnabled$.subscribe(autoEnabled => {
-                if (autoEnabled && this.needsTransaction) {
-                    this.needsTransaction = false;
-                    this.loadData();
-                }
-            })
-        );
-    }
-
-    ngOnDestroy() {
-        this.subscriptions.forEach(sub => sub.unsubscribe());
-    }
-
-    /** Returns true if we're in manual mode without an open transaction */
-    private checkNeedsTransaction(): boolean {
-        if (!this.driver.autoTransactionEnabled$.value && !this.driver.transactionOpen) {
-            this.needsTransaction = true;
-            return true;
-        }
-        return false;
-    }
-
-    /** Called from template to open a transaction and load data */
-    openTransactionAndLoad() {
-        this.driver.openTransaction("read").subscribe();
-        // Data will load automatically via the transaction$ subscription
-    }
-
-    /** Refresh data with current transaction */
-    refresh() {
-        this.loadData();
-    }
-
-    private loadData() {
-        // Record current transaction ID for stale detection
-        this.loadedWithTransactionId = this.driver.currentTransaction?.id ?? null;
-        this.isDataStale = false;
-
-        if (this.type.kind === "attributeType") {
-            // For attributes, fetch owners instead of attributes/relations
-            this.fetchOwners();
-        } else {
-            this.fetchAttributes();
-            this.fetchRelations();
-            if (this.type.kind === "relationType") {
-                this.fetchOwnLinks();
-            }
-        }
-    }
-
-    private fetchAttributes() {
-        // Check if type has any owned attributes (only entities and relations can own attributes)
-        const hasOwnedAttributes = "ownedAttributes" in this.type && this.type.ownedAttributes.length > 0;
-
-        if (!hasOwnedAttributes) {
-            // No attributes to fetch - nothing to do
-            this.loading = false;
-            this.attributes = [];
-            return;
-        }
-
-        this.loading = true;
-
-        // Fetch all attributes for this instance
-        const query = `
-match
-    $instance isa ${this.type.label};
-    $instance iid ${this.instanceIID};
-    $instance has $attr;
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                this.loading = false;
-
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching attributes: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processAttributes(answers);
-                }
-            },
-            error: (err) => {
-                this.loading = false;
-                this.snackbar.errorPersistent(`Error fetching attributes: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private processAttributes(conceptRowAnswers: ConceptRowAnswer[]) {
-        // Group attributes by type
-        const attrMap = new Map<string, { valueType: string; values: string[] }>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const attrConcept = row["attr"];
-            if (attrConcept && attrConcept.kind === "attribute") {
-                const attrType = attrConcept.type?.label || "unknown";
-                const attrValueType = attrConcept.type?.valueType || "unknown";
-                const attrValue = String(attrConcept.value);
-
-                if (!attrMap.has(attrType)) {
-                    attrMap.set(attrType, { valueType: attrValueType, values: [] });
-                }
-                attrMap.get(attrType)!.values.push(attrValue);
-            }
-        }
-
-        // Convert to AttributeData array
-        this.attributes = Array.from(attrMap.entries()).map(([type, data]) => ({
-            type,
-            valueType: data.valueType,
-            values: data.values,
-        }));
-    }
-
-    private typeOrSubtypePlaysRoles(type: SchemaConcept): boolean {
-        if ("playedRoles" in type && type.playedRoles.length > 0) return true;
-        if ("subtypes" in type) {
-            for (const sub of type.subtypes) {
-                if (this.typeOrSubtypePlaysRoles(sub)) return true;
-            }
-        }
-        return false;
-    }
-
-    private fetchRelations() {
-        // Skip if the type (and all subtypes) play no roles in any relation
-        if (!this.typeOrSubtypePlaysRoles(this.type)) {
-            this.relationsLoading = false;
-            this.allRelations = [];
-            return;
-        }
-
-        this.relationsLoading = true;
-
-        // Fetch all relations where this instance plays a role, along with all links
-        const query = `
-match
-    $instance iid ${this.instanceIID};
-    $rel links ($role: $instance);
-    $rel links ($otherRole: $player);
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                this.relationsLoading = false;
-
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching relations: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processRelations(answers);
-                }
-            },
-            error: (err) => {
-                this.relationsLoading = false;
-                this.snackbar.errorPersistent(`Error fetching relations: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private fetchOwnLinks() {
-        this.linksLoading = true;
-
-        // Fetch links of this relation instance
-        const query = `
-match
-    $rel iid ${this.instanceIID};
-    $rel links ($role: $player);
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                this.linksLoading = false;
-
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching links: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processOwnLinks(answers);
-                }
-            },
-            error: (err) => {
-                this.linksLoading = false;
-                this.snackbar.errorPersistent(`Error fetching links: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private processOwnLinks(conceptRowAnswers: ConceptRowAnswer[]) {
-        const linkMap = new Map<string, LinkData>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const role = row["role"];
-            const player = row["player"];
-
-            if (!isRoleType(role)) {
-                throw new Error(`Expected role type for $role, got: ${JSON.stringify(role)}`);
-            }
-            if (!isInstance(player)) {
-                throw new Error(`Expected entity or relation for $player, got: ${JSON.stringify(player)}`);
-            }
-
-            const roleLabel = role.label;
-            const playerIID = player.iid;
-            const playerTypeLabel = player.type.label;
-
-            // Use role+playerIID as key to avoid duplicates
-            const key = `${roleLabel}:${playerIID}`;
-            if (!linkMap.has(key)) {
-                linkMap.set(key, {
-                    roleLabel,
-                    playerIID,
-                    playerTypeLabel,
-                    isSelf: false,
-                    attributes: [],
-                    expanded: false,
-                });
-            }
-        }
-
-        this.ownLinks = Array.from(linkMap.values());
-        this.loadOwnLinkAttributes();
-    }
-
-    private loadOwnLinkAttributes() {
-        if (this.ownLinks.length === 0) return;
-
-        const iidList = this.ownLinks.map(rp => rp.playerIID);
-        const iidConditions = iidList.map(iid => `{ $player iid ${iid}; }`).join(" or ");
-        const query = `
-match
-    ${iidConditions};
-    $player has $attr;
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching link attributes: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    const playerAttrMap = this.buildPlayerAttributeMap(answers);
-
-                    for (const link of this.ownLinks) {
-                        const attrMap = playerAttrMap.get(link.playerIID);
-                        if (attrMap) {
-                            link.attributes = Array.from(attrMap.entries()).map(([type, data]) => ({
-                                type,
-                                valueType: data.valueType,
-                                values: data.values,
-                            }));
-                        }
-                    }
-                }
-            },
-            error: (err) => {
-                this.snackbar.errorPersistent(`Error fetching link attributes: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private processRelations(conceptRowAnswers: ConceptRowAnswer[]) {
-        // Build flat list of relation instances, keyed by IID
-        const instanceMap = new Map<string, RelationInstanceData>();
-        const typeSet = new Set<string>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const rel = row["rel"];
-            const otherRole = row["otherRole"];
-            const player = row["player"];
-
-            if (!rel || rel.kind !== "relation") {
-                throw new Error(`Expected relation for $rel, got: ${JSON.stringify(rel)}`);
-            }
-            if (!isInstance(player)) {
-                throw new Error(`Expected entity or relation for $player, got: ${JSON.stringify(player)}`);
-            }
-            if (!isRoleType(otherRole)) {
-                throw new Error(`Expected role type for $otherRole, got: ${JSON.stringify(otherRole)}`);
-            }
-
-            const relTypeLabel = rel.type.label;
-            const relIID = rel.iid;
-            const roleLabel = otherRole.label;
-            const playerIID = player.iid;
-            const playerTypeLabel = player.type.label;
-
-            typeSet.add(relTypeLabel);
-
-            // Initialize relation instance if needed
-            if (!instanceMap.has(relIID)) {
-                instanceMap.set(relIID, {
-                    relationTypeLabel: relTypeLabel,
-                    relationIID: relIID,
-                    links: [],
-                    attributes: [],
-                    attributesLoaded: false,
-                });
-            }
-            const relInstance = instanceMap.get(relIID)!;
-
-            // Add link if not already present
-            const isSelf = playerIID === this.instanceIID;
-            const existingLink = relInstance.links.find(
-                rp => rp.playerIID === playerIID && rp.roleLabel === roleLabel
-            );
-            if (!existingLink) {
-                relInstance.links.push({
-                    roleLabel,
-                    playerIID,
-                    playerTypeLabel,
-                    isSelf,
-                    attributes: [],
-                    expanded: false,
-                });
-            }
-        }
-
-        // Store flat list and unique types
-        // Sort links so "this" (self) always appears last
-        const relations = Array.from(instanceMap.values());
-        for (const rel of relations) {
-            rel.links.sort((a, b) => (a.isSelf === b.isSelf ? 0 : a.isSelf ? 1 : -1));
-        }
-        this.allRelations = relations;
-        this.relationTypes = Array.from(typeSet).sort();
-
-        // Load all link attributes and relation attributes upfront
-        this.loadAllLinkAttributes();
-        this.loadAllRelationAttributes();
-    }
-
     openLinkDetail(link: LinkData, event: Event) {
         event.stopPropagation();
-        const schema = this.schemaState.value$.value;
-        if (!schema) {
-            this.snackbar.errorPersistent("Schema not loaded");
-            return;
-        }
+        this.navigateToInstance(link.playerTypeLabel, link.playerIID);
+    }
 
-        // Find the type in schema (could be entity or relation)
-        const playerType = schema.entities[link.playerTypeLabel] || schema.relations[link.playerTypeLabel];
-        if (!playerType) {
-            this.snackbar.errorPersistent(`Type '${link.playerTypeLabel}' not found in schema`);
-            return;
-        }
-
-        // Build new breadcrumbs by appending current instance
-        const newBreadcrumbs: BreadcrumbItem[] = [
-            ...this.breadcrumbs,
-            { kind: "instance-detail", typeLabel: this.type.label, typeKind: this.type.kind, instanceIID: this.instanceIID }
-        ];
-
-        this.dataEditorState.openInstanceDetail(playerType, link.playerIID, newBreadcrumbs);
+    openOwnLinkDetail(link: LinkData, event: Event) {
+        event.stopPropagation();
+        this.navigateToInstance(link.playerTypeLabel, link.playerIID);
     }
 
     openRelationDetail(instance: RelationInstanceData, event: Event) {
         event.stopPropagation();
+        this.navigateToInstance(instance.relationTypeLabel, instance.relationIID);
+    }
+
+    openOwnerDetail(owner: OwnerData, event: Event) {
+        event.stopPropagation();
+        this.navigateToInstance(owner.ownerTypeLabel, owner.ownerIID);
+    }
+
+    private navigateToInstance(typeLabel: string, instanceIID: string) {
         const schema = this.schemaState.value$.value;
         if (!schema) {
             this.snackbar.errorPersistent("Schema not loaded");
             return;
         }
-
-        const relationType = schema.relations[instance.relationTypeLabel];
-        if (!relationType) {
-            this.snackbar.errorPersistent(`Relation type '${instance.relationTypeLabel}' not found in schema`);
+        const targetType = schema.entities[typeLabel] || schema.relations[typeLabel];
+        if (!targetType) {
+            this.snackbar.errorPersistent(`Type '${typeLabel}' not found in schema`);
             return;
         }
-
-        // Build new breadcrumbs by appending current instance
         const newBreadcrumbs: BreadcrumbItem[] = [
             ...this.breadcrumbs,
             { kind: "instance-detail", typeLabel: this.type.label, typeKind: this.type.kind, instanceIID: this.instanceIID }
         ];
-
-        this.dataEditorState.openInstanceDetail(relationType, instance.relationIID, newBreadcrumbs);
+        this.dataEditorState.openInstanceDetail(targetType, instanceIID, newBreadcrumbs);
     }
 
     navigateToBreadcrumb(breadcrumb: BreadcrumbItem, index: number) {
@@ -562,39 +109,7 @@ match
     }
 
     exploreType() {
-        // Open type table with empty breadcrumbs (fresh start)
         this.dataEditorState.openTypeTab(this.type, [], true);
-    }
-
-    openOwnLinkDetail(link: LinkData, event: Event) {
-        event.stopPropagation();
-        const schema = this.schemaState.value$.value;
-        if (!schema) {
-            this.snackbar.errorPersistent("Schema not loaded");
-            return;
-        }
-
-        const playerType = schema.entities[link.playerTypeLabel] || schema.relations[link.playerTypeLabel];
-        if (!playerType) {
-            this.snackbar.errorPersistent(`Type '${link.playerTypeLabel}' not found in schema`);
-            return;
-        }
-
-        const newBreadcrumbs: BreadcrumbItem[] = [
-            ...this.breadcrumbs,
-            { kind: "instance-detail", typeLabel: this.type.label, typeKind: this.type.kind, instanceIID: this.instanceIID }
-        ];
-
-        this.dataEditorState.openInstanceDetail(playerType, link.playerIID, newBreadcrumbs);
-    }
-
-    get typeKindLabel(): string {
-        switch (this.type.kind) {
-            case "entityType": return "entity";
-            case "relationType": return "relation";
-            case "attributeType": return "attribute";
-            default: return "";
-        }
     }
 
     getTypeIconClass(typeKind: string, filled: boolean): string {
@@ -609,380 +124,5 @@ match
             default:
                 return `${style} fa-question`;
         }
-    }
-
-    private formatAttributeValue(value: string, valueType: string): string {
-        switch (valueType) {
-            case "string":
-                // Escape quotes and wrap in quotes
-                return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
-            case "boolean":
-            case "long":
-            case "double":
-                // Numeric and boolean values don't need quotes
-                return value;
-            case "datetime":
-            case "datetime-tz":
-                // DateTime values are returned as ISO strings
-                return value;
-            case "date":
-                // Date values
-                return value;
-            case "duration":
-                // Duration values
-                return value;
-            default:
-                // Default to quoting as string
-                return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
-        }
-    }
-
-    private fetchOwners() {
-        this.ownersLoading = true;
-
-        // For attributes, instanceIID is actually the attribute value (not a real IID)
-        // We need to query by value, formatting appropriately for the value type
-        const attrType = this.type as SchemaAttribute;
-        const valueType = attrType.valueType;
-        const formattedValue = this.formatAttributeValue(this.instanceIID, valueType);
-
-        // Fetch all entities/relations that own this attribute
-        const query = `
-match
-    $owner has ${this.type.label} ${formattedValue};
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                this.ownersLoading = false;
-
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching owners: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processOwners(answers);
-                }
-            },
-            error: (err) => {
-                this.ownersLoading = false;
-                this.snackbar.errorPersistent(`Error fetching owners: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private processOwners(conceptRowAnswers: ConceptRowAnswer[]) {
-        const ownerMap = new Map<string, OwnerData>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const owner = row["owner"];
-
-            if (!isInstance(owner)) {
-                throw new Error(`Expected entity or relation for $owner, got: ${JSON.stringify(owner)}`);
-            }
-
-            const ownerIID = owner.iid;
-            if (!ownerMap.has(ownerIID)) {
-                ownerMap.set(ownerIID, {
-                    ownerIID,
-                    ownerTypeLabel: owner.type.label,
-                    ownerKind: owner.kind as "entity" | "relation",
-                    attributes: [],
-                });
-            }
-        }
-
-        this.owners = Array.from(ownerMap.values());
-        this.loadOwnerAttributes();
-    }
-
-    private loadOwnerAttributes() {
-        if (this.owners.length === 0) return;
-
-        const iidList = this.owners.map(o => o.ownerIID);
-        const iidConditions = iidList.map(iid => `{ $owner iid ${iid}; }`).join(" or ");
-        const query = `
-match
-    ${iidConditions};
-    $owner has $attr;
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching owner attributes: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processOwnerAttributes(answers);
-                }
-            },
-            error: (err) => {
-                this.snackbar.errorPersistent(`Error fetching owner attributes: ${extractErrorMessage(err)}`);
-            }
-        });
-    }
-
-    private processOwnerAttributes(conceptRowAnswers: ConceptRowAnswer[]) {
-        // Build a map of owner IID -> attribute type -> { valueType, values }
-        const ownerAttrMap = new Map<string, Map<string, { valueType: string; values: string[] }>>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const owner = row["owner"];
-            const attr = row["attr"];
-
-            if (!isInstance(owner)) {
-                throw new Error(`Expected entity or relation for $owner, got: ${JSON.stringify(owner)}`);
-            }
-            if (!attr || attr.kind !== "attribute") {
-                throw new Error(`Expected attribute for $attr, got: ${JSON.stringify(attr)}`);
-            }
-
-            const ownerIID = owner.iid;
-            const attrType = attr.type.label;
-            const attrValueType = attr.type.valueType || "unknown";
-            const attrValue = String(attr.value);
-
-            if (!ownerAttrMap.has(ownerIID)) {
-                ownerAttrMap.set(ownerIID, new Map());
-            }
-            const attrMap = ownerAttrMap.get(ownerIID)!;
-
-            if (!attrMap.has(attrType)) {
-                attrMap.set(attrType, { valueType: attrValueType, values: [] });
-            }
-            attrMap.get(attrType)!.values.push(attrValue);
-        }
-
-        // Assign attributes to owners
-        for (const owner of this.owners) {
-            const attrMap = ownerAttrMap.get(owner.ownerIID);
-            if (attrMap) {
-                owner.attributes = Array.from(attrMap.entries()).map(([type, data]) => ({
-                    type,
-                    valueType: data.valueType,
-                    values: data.values,
-                }));
-            }
-        }
-    }
-
-    openOwnerDetail(owner: OwnerData, event: Event) {
-        event.stopPropagation();
-        const schema = this.schemaState.value$.value;
-        if (!schema) {
-            this.snackbar.errorPersistent("Schema not loaded");
-            return;
-        }
-
-        const ownerType = owner.ownerKind === "entity"
-            ? schema.entities[owner.ownerTypeLabel]
-            : schema.relations[owner.ownerTypeLabel];
-
-        if (!ownerType) {
-            this.snackbar.errorPersistent(`Type '${owner.ownerTypeLabel}' not found in schema`);
-            return;
-        }
-
-        const newBreadcrumbs: BreadcrumbItem[] = [
-            ...this.breadcrumbs,
-            { kind: "instance-detail", typeLabel: this.type.label, typeKind: this.type.kind, instanceIID: this.instanceIID }
-        ];
-
-        this.dataEditorState.openInstanceDetail(ownerType, owner.ownerIID, newBreadcrumbs);
-    }
-
-    private loadAllLinkAttributes() {
-        // Collect all non-self link IIDs from all instances
-        const allIIDs = new Set<string>();
-
-        for (const instance of this.allRelations) {
-            for (const rp of instance.links) {
-                if (!rp.isSelf) {
-                    allIIDs.add(rp.playerIID);
-                }
-            }
-        }
-
-        if (allIIDs.size === 0) {
-            for (const instance of this.allRelations) {
-                instance.attributesLoaded = true;
-            }
-            return;
-        }
-
-        // Build query for all links' attributes
-        const iidList = Array.from(allIIDs);
-        const iidConditions = iidList.map(iid => `{ $player iid ${iid}; }`).join(" or ");
-        const query = `
-match
-    ${iidConditions};
-    $player has $attr;
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching link attributes: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processAllLinkAttributes(answers);
-                }
-                for (const instance of this.allRelations) {
-                    instance.attributesLoaded = true;
-                }
-            },
-            error: (err) => {
-                const message = err?.message || (typeof err === "object" ? JSON.stringify(err) : String(err));
-                this.snackbar.errorPersistent(`Error fetching link attributes: ${message}`);
-                for (const instance of this.allRelations) {
-                    instance.attributesLoaded = true;
-                }
-            }
-        });
-    }
-
-    private loadAllRelationAttributes() {
-        // Collect all relation IIDs
-        const allIIDs = this.allRelations.map(r => r.relationIID);
-
-        if (allIIDs.length === 0) {
-            return;
-        }
-
-        // Build query for all relations' attributes
-        const iidConditions = allIIDs.map(iid => `{ $rel iid ${iid}; }`).join(" or ");
-        const query = `
-match
-    ${iidConditions};
-    $rel has $attr;
-        `.trim();
-
-        this.driver.query(query).subscribe({
-            next: (res: ApiResponse<QueryResponse>) => {
-                if (isApiErrorResponse(res)) {
-                    this.snackbar.errorPersistent(`Error fetching relation attributes: ${res.err.message}`);
-                    return;
-                }
-
-                if (res.ok.answerType === "conceptRows") {
-                    const answers = (res.ok as any).answers as ConceptRowAnswer[];
-                    this.processAllRelationAttributes(answers);
-                }
-            },
-            error: (err) => {
-                const message = err?.message || (typeof err === "object" ? JSON.stringify(err) : String(err));
-                this.snackbar.errorPersistent(`Error fetching relation attributes: ${message}`);
-            }
-        });
-    }
-
-    private processAllRelationAttributes(conceptRowAnswers: ConceptRowAnswer[]) {
-        // Build a map of relation IID -> attribute type -> { valueType, values }
-        const relAttrMap = new Map<string, Map<string, { valueType: string; values: string[] }>>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const rel = row["rel"];
-            const attr = row["attr"];
-
-            if (!rel || rel.kind !== "relation") {
-                throw new Error(`Expected relation for $rel, got: ${JSON.stringify(rel)}`);
-            }
-            if (!attr || attr.kind !== "attribute") {
-                throw new Error(`Expected attribute for $attr, got: ${JSON.stringify(attr)}`);
-            }
-
-            const relIID = rel.iid;
-            const attrType = attr.type.label;
-            const attrValueType = attr.type.valueType || "unknown";
-            const attrValue = String(attr.value);
-
-            if (!relAttrMap.has(relIID)) {
-                relAttrMap.set(relIID, new Map());
-            }
-            const attrMap = relAttrMap.get(relIID)!;
-
-            if (!attrMap.has(attrType)) {
-                attrMap.set(attrType, { valueType: attrValueType, values: [] });
-            }
-            attrMap.get(attrType)!.values.push(attrValue);
-        }
-
-        // Assign attributes to all relations
-        for (const instance of this.allRelations) {
-            const attrMap = relAttrMap.get(instance.relationIID);
-            if (attrMap) {
-                instance.attributes = Array.from(attrMap.entries()).map(([type, data]) => ({
-                    type,
-                    valueType: data.valueType,
-                    values: data.values,
-                }));
-            }
-        }
-    }
-
-    private processAllLinkAttributes(conceptRowAnswers: ConceptRowAnswer[]) {
-        // Build a map of player IID -> attribute type -> { valueType, values }
-        const playerAttrMap = this.buildPlayerAttributeMap(conceptRowAnswers);
-
-        // Assign attributes to all links across all instances
-        for (const instance of this.allRelations) {
-            for (const link of instance.links) {
-                if (link.isSelf) continue;
-
-                const attrMap = playerAttrMap.get(link.playerIID);
-                if (attrMap) {
-                    link.attributes = Array.from(attrMap.entries()).map(([type, data]) => ({
-                        type,
-                        valueType: data.valueType,
-                        values: data.values,
-                    }));
-                }
-            }
-        }
-    }
-
-    private buildPlayerAttributeMap(conceptRowAnswers: ConceptRowAnswer[]): Map<string, Map<string, { valueType: string; values: string[] }>> {
-        const playerAttrMap = new Map<string, Map<string, { valueType: string; values: string[] }>>();
-
-        for (const answer of conceptRowAnswers) {
-            const row = answer.data;
-            const player = row["player"];
-            const attr = row["attr"];
-
-            if (!isInstance(player)) {
-                throw new Error(`Expected entity or relation for $player, got: ${JSON.stringify(player)}`);
-            }
-            if (!attr || attr.kind !== "attribute") {
-                throw new Error(`Expected attribute for $attr, got: ${JSON.stringify(attr)}`);
-            }
-
-            const playerIID = player.iid;
-            const attrType = attr.type.label;
-            const attrValueType = attr.type.valueType || "unknown";
-            const attrValue = String(attr.value);
-
-            if (!playerAttrMap.has(playerIID)) {
-                playerAttrMap.set(playerIID, new Map());
-            }
-            const attrMap = playerAttrMap.get(playerIID)!;
-
-            if (!attrMap.has(attrType)) {
-                attrMap.set(attrType, { valueType: attrValueType, values: [] });
-            }
-            attrMap.get(attrType)!.values.push(attrValue);
-        }
-
-        return playerAttrMap;
     }
 }
